Reverted some recent changes.
I got carried away and wanted to use some of Qt Designer's components,
but this is not the right approach. At least two reasons:
    1) we need three columns (parameter, value and units) and not only
two; and
    2) Qt Designer's property editor displays the properties of a
QObject instance. So, we would have to dynamically set properties to an
'empty' QObject instance, but then there is the issue of the
'objectName' property being there by default… Also, what about
properties having properties of their own?

Whatever the case, it seems like Qt Designer's property editor is not
the way to go. Instead, we will probably have to use a QTreeView.