A stupidily HUGE commit regarding our CellML-related code. Note that, at this stage, the CellML annotation plugin doesn't work anymore. We need to do some more work on our new RDF code.
OpenCOR doesn't use its own internal representation of a CellML file
anymore. It was just going to be too cumbersome in the long run. The
only thing we are keeping is the RDF stuff since we did add quite a bit
of a layer on top of what the CellML API offers. In the process, we
also improved our use of ObjRef.