QTreeViewWidget: removed the mouse double-click handling.
We were originally trying to address an issue with double-clicking on a
checkbox which would expand/collapse a graph property (in the
simulation view) and that was because of our handling a mouse
double-click. It was probably handled because of something else
preventing a mouse double-click from working? Whatever the case, mouse
double-clicking expands/collapses a property unless it's done over a
check box, in which case it will just change the check box state.