Some work on our Travis CI script.
Removed Clang as one of the compilers since our binaries for the CellML
API and LLVM were built using gcc, which means that we cannot build
OpenCOR with Clang on Ubuntu. Also limited the number of parallel jobs
for the compilation. We used to have no limit and this would eventually
crash the build. Oops!