Reimplemented the need to specify whether a plugin is selectable.
Indeed, if only one plugin was to be available (e.g. the Core plugin),
then it used to be selectable no matter what while only a truly
selectable plugin (e.g. the RawView plugin) should be made selectable.
It’s because we used to determine the selectable state of a plugin
based on whether it was needed by another plugin.