Travis-CI: use the OS's default compiler.
Indeed, to build a release version of OpenCOR using Clang on Linux
fails. There is an issue with __extern_always_inline being unknown
while it’s all fine when building a debug version. Anyway, no time to
investigate, not to mention that we always use the default compiler on
a given OS, so…