CellML support: beautify the serialisation of a CellML file (closes #497).
The difference with the previous solution is that now we ensure the
order of the attributes. When it comes to XML, it doesn’t matter
whether an attribute comes before another or not. However, to have
attributes sorted makes it for easier comparison, which can be quite
handy at times.