Rely on JSON files to download the different versions/snapshots of OpenCOR (closes #505).
Those are not pure JSON files anymore, but JavaScript files that
contain some JSON. Indeed, to have pure JSON files prevented us from
testing the download pages locally. The only way to test them was
through a web server, which was exactly what we want to avoid (see
issue #490). The current solution still allows us to do what we need
for issue #1, so we are fine. It’s just not as ‘neat’…