Improved our way to check that a file is a CellML file [ci skip].
In the case of a file that didn’t have the expected .cellml file
extension, we used to load the file as an XML file and check whether
its root element was a model element with either the CellML 1.0 or 1.1
namespace. Now, we simply try to load the file using the CellML API,
which is better for the future (e.g. we will automatically handle
future versions of CellML).