Pretty CellML view: some work on providing syntax highlighting for the pretty CellML format (#530) [ci skip].
Before, we used to split the text to style into lines, but that caused
problems when styling parameters that didn’t fit one line. Also, this
would have caused problems for comments such as /* I am a comment */.
So, now, we style the whole line recursively. At this stage, we can
style comments that use //.