Reverted some changes related to the loading of a CellML file using namespace processing.
Indeed, it would seem that if a model element has a cmeta:id and that
it listed before the xmlns:cmeta attribute, then the cmeta:id won’t be
recognised when trying to retrieve it using
pDomNode.attributes().namedItemNS(CellMLSupport::CmetaIdNamespace,
"id”). If we were to put the xmlns:cmeta attribute before the cmeta:id,
then it’s all fine. However, my understanding is that the order of
attributes doesn’t matter, yet it seems that it does in this particular
case when using QDomDocument & Co.