Some more reverting [ci skip].
I don’t know why, but I kind of assumed that we didn’t need to clean a
MathML block. Well, wrong! Our MathML viewer indirectly crashes if a
MathML block isn’t cleaned (e.g. when we have a cn element with a
cellml:units attribute. If the attribute is left in, then we get a
crash!