Tests: removed our QCOMPARELIST macro.
It proved to be very slow for some tests (parsing/conversion of/from
the CellML Text format). It could have been improved, but in the end
the default comparison of QStringList objects seems sufficient when
lists don’t have the same contents (despite what I seemed to remember;
could that have been improved in Qt 5.4.1?).