Single Cell view: very very tiny improvement to the accuracy of our timer [ci skip].
This is only relevant when we are running a simulation, pausing it and resuming it. Indeed, we used to restart the timer, but it's the same as starting it, except that QElapsedTimer::restart() also returns the time elapsed since the previous start. We don't need that information, so we might as well simply call QElapsedTimer::start()...