Fixed an issue with OpenCOR crashing on Mac OS X 10.7 and OS X 10.8 when a model needs to compute 10^x (closes #611) [ci skip].
To use i686-pc-win32-elf and x86_64-pc-win32-elf for Win32 and Win64, respectively, is not good enough. Indeed, this works fine on Win64, but on Win32 we get the same problem as the one reported in issue #583. So, instead, we now use i686-pc-windows-msvc-elf and x86_64-pc-windows-msvc-elf, which we know work fine (thanks to the LLVM patch used to fix issue #583).