Travis CI: trying to understand why the CellML Tools tests now fail on Linux.
Ok, everything works fine using the master branch, so it’s not an issue
with our prebuilt libgit2, but what could it be then since it’s working
fine if we use a non-prebuilt libgit2?

Trying to move the tracking of a plugin’s full dependencies further up.
Could that somehow solve the problem?

Also, FWIW, Travis CI runs on Ubuntu 12.04 LTS while we support 14.04
LTS and later. Still, as previously said, it all works fine with the
master branch, so it shouldn’t matter as such.