Allow Core::getSaveFileName() to return the same file name [ci skip].
The original idea was that if we were to save a file under a new name,
then we couldn’t use the same name for the file. However, this can be a
problem if we want to export a SED-ML file since we will give
Core::getSaveFileName() the file path of a SED-ML. So, if we were to
agree to using that file name, Core::getSaveFileName() would complain,
which is not what we want. So, instead, it should be the caller who
should handle, if needed, the case where we are using the same file
name.