ZIP support: make sure that we can extract files in subdirectories (#757).
Say that you want to add ```dir/file.txt``` to a ZIP writer. Then, to
extract all of the ZIP file will fail because ```dir``` doesn’t exist.
Now, I would have thought that this could have been solved by adding
the ```dir``` directory to our ZIP writer, but this then results in a
ZIP file that cannot be unzipped using OS X’s default unzipper. So,
instead, during the extraction, we check that the folder in which a
file should be extracted exists and if it doesn’t then create it.