Viewer widget: improved the rendering of certain equations [ci skip].
To render (with subscripts on)

    a = -b_c*d_e+f;

used to result in b_c being rendered with c as a subscript of b (as
expected), but then we would have expected the same for d_e (i.e. have
e rendered as a subscript of d), but instead it was simply rendered as
d_e (i.e. no subscript whatsoever).