Kind of reverting commit 59f3ce3 (related to issue #868) [ci skip].
Not only is it not needed anymore because of
SingleCellViewGraphPanelPlotWidget::axesChanged() and
SingleCellViewGraphPanelPlotWidget::resetAxesChanged() now being
available, but our ‘fix’ resulted in the loop always starting at 0
(!!), meaning that we were always testing all of the data rather than
just the new one! In other words, the longer the simulation the more
rugged the rendering.