Plugin resources: rename them to address a Linux issue [ci skip].
Ideally, we would have our resource files named i18n.qrc.in and ui.qrc
for all our plugins, but this is causing problems on Linux with Qt
apparently accepting only one file called i18n.qrc.in or ui.qrc. The
end result is that, on a development machine, plugin resources (e.g.
icons, translations) are not available while, strangely enough, they
are on a deployment machine...