Busy widget: remove the semi-transparency option [ci skip].
It just doesn’t look ‘right’ from a GUI viewpoint. At the end of the
day, we still have what we wanted in the first place, i.e. have the
parent really unusable. Indeed, even though, in the PMR window, we do
disable the web viewer, that viewer can still be used. However, now
with the busy widget covering all of its parent (and using a
transparent background for the non-busy part of its rendering), we are
fine (i.e. the parent can really be considered ‘disabled’).