Standard file manager: fixed a problem with wrongly deleting some file objects.
Indeed, our StandardFileManager class would delete all the objects held
by mFiles, but it shouldn’t have. People should just ensure that things
are being unmanaged.

The end result is that when we were, for example, trying to export a
CellML file to C from the command, the export itself would work fine,
but then OpenCOR would crash on exit…