Qt: some work towards upgrading to version 5.5.1 (#971).
Addresses the issue about the Single Cell view: to simulate the N62
model with one or several graphs and with, for example, a 200,000 slow
down results in the plotting area repeatedly becoming black.

This time, we have a proper fix. Indeed, to disable the temporary event
filter wasn’t the right approach since it resulted in the whole
plotting area being repainted. It was ‘fine’ on my development machine
(OS X 10.11), but not on a vanilla OS X 10.8-10.10 machine where
plotting 1,000,000 data points resulted in chunks of data points being
plotted at once…

The ‘new’ fix consists of setting the CopyBackingStore attribute to
true. This has a slight disadvantage which is that the backing store is
constantly copied over to the canvas, but at least it works as expected
on both my development machine and OS X 10.8-10.10.