Addressed an issue with deploying OpenCOR on Ubuntu 16.04 LTS [ci skip].
Although we did originally add support for Ubuntu 16.04 LTS (see issue
#965), the fact that we now build Qt WebKit ourselves (see issue #1015)
means that our copy of Qt WebKit (somehow) uses the system version of
ICU. Since we build Qt WebKit on Ubuntu 14.04 LTS, it means that
deployment on that platform was fine, but not on Ubuntu 16.04 LTS
(since the system version of ICU on that platform is different). So,
the solution, here, is ‘simply’ to deploy both the system version of
ICU on which Qt WebKit depends, as well as the Qt version of ICU used
by the other Qt libraries we got from Qt itself.