Linux: ship the standard C++ library used to build Qt WebKit.
Indeed, to address issue #1037, we had to upgrade to GCC/G++ 4.9+,
which means using a different version of libstdc++ than the one shipped
with Ubuntu 14.04 LTS (while it is available on Ubuntu 16.04 LTS). So,
we need to ship it with OpenCOR in case the version of Ubuntu used by
someone is not recent enough.