CMake: make sure that we can properly find NSIS (when available).
Indeed, up to CMake 3.5.x, CMake was only avaialble as a 32-bit application on Windows. However, starting with CMake 3.6.x, CMake is now available both as a 32-bit and 64-bit application.

Now, depending on the version you use, the value of `$ENV{PROGRAMFILES}` will return either `C:\Program Files (x86)` (when using the 32-bit version of CMake) or `C:\Program Files` (when using the 64-bit version of CMake), while `$ENV{PROGRAMW6432}` will always return `C:\Program Files` in both cases.

There is therefore something funny going on with the 64-bit version of CMake. So, to be on the safe side, we 'manually' look for the possible location of NSIS.