Travis CI: done determining why things were not working on Linux anymore.
On Linux, to link CellMLTextView_conversiontests and
CellMLTextView_parsingtests results in an error about
`__cxa_throw_bad_array_new_length@Qt_5` not being defined in
libQScintilla.so.

It’s not clear yet why it is the case since everything works fine
locally (and when deployed). However, to build QScintilla ourselves
(rather than using its prebuilt version) ‘solves’ the problem, so…