Graph panel widget: improved the plotting of data (related to issue #671).
I thought I was done with issue #671, but I got a black plot when
trying to plot two successive (in the list of parameters)  parameters
against the VOI. Upon analysis, the black plot was due to
GraphPanelPlotWidget::drawGraphFrom() replotting itself and reallowing
direct painting. However, the replotting wasn’t done yet and because
direct painting was reallowed, the next call to
GraphPanelPlotWidget::drawGraphFrom() (for the second model parameter)
did a direct painting. Hence, the black plot.

The idea was therefore not reallowing direct painting straightaway, but
only after the replotting has occurred. While at it, we also prevent
successive calls to replot() in GraphPanelPlotWidget::drawGraphFrom().