General: fixed a problem with some plugins potentially not loading on some Linux boxes (due to some RPATH/RUNPATH values) (closes #1107).
Argh, to set the RPATH of a library that is already correct changes its
SHA-1 value, which means that every time we build OpenCOR on Linux, we
need to redownload all of our prebuilt plugins. So, now, we check the
RPATH value and compare it against the one we want to use.