Reverting commit 8bca5dd78d64b1133b69d109434ee688413ee5c5 [ci skip].
Indeed, although it does allow QScintilla to be compiled using Xcode 8,
if you open more than one file using the `CellML Text` view, then the
lexer of the file will be different from the others on Linux and macOS
(at least, using our current lexer settings for the `CellML Text`
view). Indeed, the height of a line is a slightly bigger for the first
editor, which means that when you switch to another file, you can see
that a slight ‘shift up’ in the text.