QScintilla: fix Xcode 8 compilation problem.
Another solution has been given on the QScintilla mailing list (see
https://www.riverbankcomputing.com/pipermail/qscintilla/2016-October/001
160.html), referencing
https://github.com/sqlitebrowser/sqlitebrowser/commit/969e263e5bf71e63df
8ee908cb4a2014b9a55c9b, but it’s not working for us.

Indeed, although it does allow QScintilla to be compiled using Xcode 8,
if you open more than one file using a QScintilla-based editor that
should have the same lexer, then the lexer of the first file will be
'wrong'. In our case, the line height of the first editor is slightly
bigger than that of the two other editors, which is clearly wrong.
Using this fix, everything is fine.