PropertyEditor widget: always create a value field even for a section item.
Starting with commit 467830c52f3db2ef473d6f5a47e1c4969c1ab989, we
thought it was a waste of memory to create a value field for a section
item and, in some way, it is, but the fact is that it makes things more
difficult when it comes to selecting a property or a section item.
Indeed, when selecting a property, the property value gets correctly
selected, but then if you select a section item, it will reset the
position of horizontal scrollbar, if present (even if you try to
preserve its position after selecting the section item).