PMR window: fixed a small problem when trying to directly clone a workspace [ci skip].
Indeed, the name of the directory where the cloning is to take place
wasn’t passed, which means that it was failing. Note that this didn’t
happen if we were first showing exposure files and then trying to clone
a workspace since the cloning would have been done through a different
route (PmrWindowWindow::cloneWorkspace() rather than
PmrWebService::finished() with WorkspaceInformation as a request).