PMR Workspaces window: some work on making PR #1124 more in line with the rest of OpenCOR (#1141).
Also it’s nice to see the progress of, say, cloning a workspace, there
are cases where we can’t get the progress of an action (e.g. retrieve a
user’s workspaces). So, to avoid having to determine when we should be
having a progress busy widget or a simple one, as well as handling
different types of busy widgets in the PMR and PMR Workspaces window,
we now only use a simple busy widget.