Some minor cleaning up [ci skip].
Well, in the end, it’s better to use QWebView::load()/reload() unless
QWebView::setUrl() is really necessary (e.g. in
HelpWindowWidget::retranslateUi()).

Indeed, if we think of the Help window, to go to a given page and then
go back to the home page by clicking the home page button will result
in the Help window flashing if we were to use QWebView::setUrl() while
there is none when using QWebView::load().