WebViewerWidget: some work on refactoring the code from HelpWindowWidget (#1156) [ci skip].
Note that we don’t want, in its constructor, WebViewerWidget to go to
its home page. Indeed, if we were to do that, then our Help window
wouldn’t have the state of its back button updated when clicking on a
new Help page (e.g. the “What is new?” page). This is because
HelpWindowWidget would already have about:blank has a back item in its
history. So, instead, it should be a WebViewWidget-derived class that
should go to the home page, if needed (as is now the case with
WebBrowserWindowWidget, for example).