Reverting commit dccf757 since it may cause other problems.
Indeed, our PMR support plugin relies on Settings to save its settings,
so say that you start OpenCOR and authenticate to PMR, then upon
leaving OpenCOR, your authentication information to PMR will be lost.
Now, we might be able to retrieve that information prior to saving our
settings, but it might require a new method to our Plugin interface
since no Plugin interface method is called before calling
saveSettings(). Anyway, it’s not worth it in the end and it’s better to
‘manually’ clear whatever settings needs clearing before saving our
settings.

Also,  commit dccf757 might have been fine for the GUI side of things,
but what about the CLI version of OpenCOR? In fact, we forgot to clear
the settings (as in commit dccf757) before saving them in the CLI
version, but if we had cleared them then we would have likely lost
useful GUI settings in the process.

So, yes, a ‘good’ idea gone wrong…