PMR Workspaces window: some work on our new GUI implementation.
To clear our model just before (re)populating it is critical since
before that we do some stuff with our PMR service that may require
emitting signals. In other words, to clear our model as we used to
resulted in our tree view showing up as empty for a split second when
refreshing our workspaces.