PMR Workspaces window: fixed a problem with the log on/off icon not showing properly on a Mac with a HiDPI screen resolution (closes #1204).
We used to use `QWidget::setGraphicsEffect()` to make the icon
greenish/redish and it was working on non-HiDPI resolutions, but for
some reasons things were going awry on HiDPI resolutions. So, now, we
instead rely on `QGraphicsItem::setGraphicsEffect()`. It’s a bit more
cumbersome, but not sure there is another solution…