Set the locale as early as possible [ci skip].
We used to do it at the end of MainWindow::loadSettings(), but former
note #2 is not relevant anymore (since we don’t keep track of the
number of graph panels anymore in the SingleCell view) and, in fact, it
should never have been relevant in the first place. In other words, we
should have done things in s such a way that it wasn’t dependent on
whether the locale was already set or not.