PMR support: potentially addressed a rare (?) issue with authentification [ci skip].
Sometimes when authenticating, the web page isn’t fully loaded and we
are told to close the browser. Yet, we want to wait for the page to be
fully loaded (since it looks better) and we used to call
QTimer::singleShot(0, SLOT(closeBrowser())), but it looks like it
didn’t give (in some rare cases) time for the progress bar to update
itself, meaning that the browser would never close (unless we closed it
manually). To call QTimer::singleShot(1, SLOT(closeBrowser())), i.e. we
have a 1ms delay, seems to be doing the trick…