Rely on QOpenGLWindow rather than QOpenGLWidget (#15).
To rely on QOpenGLWindow rather than QOpenGLWidget has several
advantages including:
 - We dont’ have to handle the case where the OpenGL context is about
to be destroyed (due to docking/undocking our Zinc window); and
 - Support for HiDPI resolutions works fine unlike with QOpenGLWidget.
In other words, going from 1900x1200 to 1900x1200 HiDPI doesn’t result
in the Zinc scene only taking a quarter of the window. However, we
still have to handle the case of the font getting halved.