Zinc widget: back to being a 'proper' widget (embedding a QOpenGLWindow-based window) (#15).
Note: we are committing this, but… we are going to have to use another
approach since creating a window container for our
ZincWidgetOpenglWindow object results in the whole OpenCOR application
being ‘small’ when first starting it or when resetting it all. It looks
as if our various sizeHint() methods are not being taken into account.