Zinc widget: trying with QGLWidget (#15).
Although the rendering is now fine on macOS (yes, it wasn’t using
QOpenGLWindow!?), there is still the issue of the overall size of
OpenCOR. Also, although we can undock our Zinc window, we cannot dock
it back (!?). Finally, when starting OpenCOR, the Zinc window has a
black background for a split second. Bottom line: we are reverting to
using QOpenGLWidget, even though it has some drawbacks…