LLVM package: how our CMakeLists.txt file should look like (#1306).
However, for this to work, we will need to patch our copy of LLVM+Clang
so that we can generate proper shared libraries for both LLVM and
Clang. Indeed, on macOS, for example, we can generate a proper LLVM
shared library out of the box, but not a proper Clang one (!?).