LLVMClang package: only build one shared library (#1306).
We used to build both an LLVM and Clang shared library, but although it works fine on Linux and macOS, it proves crazy on Windows. Indeed, our Clang shared library would need access to loads of LLVM symbols while we only expose those that we need in OpenCOR. To expose all the symbols needed by Clang would be a lot of work, hence we now build only one shared library instead.