SimulationExperiment view: make sure that the GUI is properly initialised when (re)loading a SED-ML file.
The problem was with opening a SED-ML file with 3 graph panels, each
with a graph. Indeed, the first graph panel was properly selected, so I
thought I would change the colour of its graph by right clicking on its
colour property, but the "Select Colour..." menu item wasn't visible
(in fact, other menu items were not properly enabled/disabled either).

So, it’s not a new problem, but one that has been around since we have
added support for SED-ML, and it was related to the fact that when
adding graph panels (to get the right number of graph panels mentioned
in the SED-ML file), we were adding them without asking for them to be
active (as expected) since we wanted the first graph panel to remain
the active one. The problem with this is that the graphs panel widget
(associated with the first graph panel) wasn't the current widget
anymore. So, upon right clicking, the context menu wasn't properly
initialised since 'crtProperty' in
SimulationExperimentViewInformationGraphsWidget::propertyEditorContextMe
nu() was 'null'...