GraphPanel widget: use the same axis optimisation algorithm for linear and logarithmic axes (#1412).
We used to try to be too ‘clever’, but it wasn’t working (!!). Indeed,
say that you had an axis with values between 14.896525 and 49.408874,
then our old optimisation would tell us to use an axis range of [10;
60] while ideally it would be [14; 50], as for a linear axis. In the
end, the algorithm for optimising linear axes seems to also work for
logarithmic axes, so we are going to use that algorithm instead (unless
we eventually find out it’s not working in some cases).