Reverted commits 7486c95 and c3e5934 (#1412).
I clearly didn’t think it through. The first commit made things quite
worse when it comes to running some simulations. By default, the X axis
would have a range of [0; 1000], but if the user was to specify a
simulation that goes from 0 to 100,000, then the X axis would keep
growing as the simulation would be run… which would make things
unnecessarily slow and not graphically pleasant…