GraphPanel widget: simplified the optimisation of axes (#1412).
We used to optimise an axis' values by making sure that they would fall onto a factor of its minor step. This was working fine with linear axes, but this is next to impossible to get to work with log axes. So, now, the only 'optimisation' is to handle the case where the minimum and maximum values of an axis are the same.