Slight reworking of SimulationExperimentViewSimulationWidget::simulationPropertyChanged() (#1412).
This is so that if the user changes the point interval and that we are using a log scale, then the axes can be updated accordingly, if needed. For example, say that the X axis uses the [0; 1000] range and that we have a point interval of 0.1, then a log scale will result in the X axis actually using the [0.1; 1000] range.