GraphPanel widget: get GraphPanelPlotWidget::dataRect() and GraphPanelPlotWidget::dataLogRect() to tell us whether they are returning something meaningful (#1412).
Indeed, the problem was that when running a simulation with a big simulation delay, then the first time we would call GraphPanelPlotWidget::dataRect() or GraphPanelPlotWidget::dataLogRect(), we would end up with QRectF(minX, minY, maxX-minX, maxY-minY) where minX = maxX and minY = maxY, which when tested using QRectF::isNull() would return true while we didn't want that. Although the values result in 'null', they are actually meaningful to us.