SimulationExperiment view: make sure that our graph panel gets properly replotted when changing some of its properties (#1426).
Indeed, if we were to pan the Y axis (so that it would now range from,
say, -10 to 990) and increase the font size, then everything would be
fine. However, if we were to reset the font size to its original value,
then we would get a ‘big’ gap to the left of the Y axis labels.