Reverting commit 66250bb.
This was a bad move since we use the parent of a CommonWidget object
when showing a busy widget. In other words, the parent gets disabled,
which can be problematic if we use the ‘wrong’ parent. This is what
happened with our CellML Annotation view. After commit 66250bb, as soon
as we would type one character in the Term field, the busy widget would
kick in and the whole OpenCOR application would be disabled. In other
words, after entering one character, we had to wait for the ontological
terms to be retrieved. Not great at all!