GraphPanel widget: make sure that our graph panels are properly aligned (#1426).
Indeed, if we had one graph panel, added a title to its Y axis, added
another graph panel and also wanted to add a title to its Y axis, then
only the Y axis title of the first graph panel would be visible. That
is, until we would, say, pan the second graph panel.

This was because the QwtScaleWidget for the Y axis of the second graph
panel didn’t have its internals updated. To ‘fix’ this, we have to
subclass QwtScaleWidget, so that we could call its layoutScale() method…