GraphPanel widget: improved our removeAllGraphPanels() method (#1426).
Indeed, until now, we were removing all the graph panels, but the first
one. This was not only arbitrary, but if the graph panel had been
customised (e.g., a title had been set), then it means that it would
still be there. So, now, we add a graph panel (since we need at least
one graph panel) and then really remove all the other graph panels.