GraphPanel widget: don’t force the alignment when not necessary (#1426).
This had been modified in commit
9c3eec2b6fee47c1b93cf6d7d1a5a7a8a2668810 to ensure proper alignment
when adding a graph panel when another had a Y title. However, this was
unnecessarily realigning things and therefore slowing things down (a
lot). So, back to what we originally had. As for adding a graph panel
when another has a Y title, it looks like our various changes (as part
of issues #1426 and #1171) indirectly addresses it since everything is
still fine.