Travis CI: (temporarily?) disabling macOS builds.
Indeed, there are currently nearly 600 jobs waiting in the macOS queue,
so… it’s going to take a long time to clear. Yet, no incident is
currently being reported on macOS, so it just looks like more and more
people are trying to build on macOS and because Travis CI doesn’t have
the same resources on macOS as on Linux, it takes much longer to build
on that platform. Now, keeping all of that in mind, we happen to CI on
Windows, Linux and macOS through Jenkins, so maybe it’s just fine to
disable macOS builds on Travis CI…? The caveat, though, is that for
some reasons our Jenkins builds are only triggered when merging
something…