General: save as such when saving a file under a new name.
Until now, if we wanted to save a non-new and non-modified file under a
new name (using the “File” | “Save As…” menu), then we used to make a
simple physical copy of the file. This was fine for a CellML file, but
not always fine for a SED-ML (while still fine for a COMBINE archive)
since the relative path to its corresponding CellML file might now be
different. So, instead, we should always save a file, whether not-new
and/or not-modified.