Slight update to our View interface now (#1366).
The idea is so that we only get the relevant MIME type when saving a
file, in cases where a view supports several MIME types. Indeed, the
Simulation Experiment view can open CellML files, SED-ML files and
COMBINE archives, but when one of the three types has been opened, we
only want to be able to save under that type.