Simulation Experiment view: made the (re)loading of remote files and SED-ML files / COMBINE archives much smoother (#1492).
Now, the only remaining issue is when OpenCOR starts with a file that
is to be shown using the Simulation Experiment view. On macOS, the tool
bar has a 3D look while it should have a flat one. It’s only that view
for the first file. Indeed, if we switch to another file using the same
view, then everything’s fine…!?