Travis CI: re-disabling macOS builds.
Well, the reason it is going much better is that they are now only
accepting one macOS job at a time! Now, because we need both release
and debug builds, it means that it takes twice as long to build as it
used to, which is not ideal. So, re-disabling macOS builds… :(