Simulation Experiment view: make it possible to save a new SED-ML file.
The downside is that the SED-ML is invalid since it doesn’t contain
everything it should. So, when it comes to saving a new SED-ML file
using the Simulation Experiment view, we should just save our default
SED-ML template, i.e. as if we were to save the SED-ML file using the
Raw SED-ML or Raw Text view.