Simulation Experiment view: a simulation shouldn't be stopped by the Simulation Experiment view (#1630).
When a simulation needs to be stopped, it should be stopped before
doing anything else (e.g. reloading a file) since it’s run in its own
thread. So, anything else will result in a potential crash of OpenCOR.
This is the reason we now stop a simulation from within our central
widget.