Graph Panel widget: improved the alignment of a graph panel with its neighbours on macOS.
Indeed, when running the model in issue #1630 on macOS, the X axis of
the first graph panel was flashing for a bit before becoming "stable".
This was due to us calling xScaleWidget->setMinBorderDist(0, 0), which
unlike our call to yScaleDraw->setMinimumExtent(0.0), we don’t seem to
need to properly align our graph panels (at least, on macOS).