Stopped using Travis CI.
Indeed, we now have everything set up and working nicely on Jenkins.
Not only that, but issue #1552 is building/testing fine locally on
Jenkins, but not on Travis CI (though our Docker Ubuntu 16.04 LTS) and
I am really not sure why. I spent a full day trying to figure out, but
to no avail. So, since Jenkins is working (and also build/test on
Windows and macOS), we might as well only use that, not to mention that
we can rely ccache on Jenkins, thus making our builds faster.