Revert "Central widget: prevent a GUI glitch when changing modes."
This reverts commit 771708d80c0d78d4765319a2fc0f6d57ef0a69d3.

Indeed, there is definitely no need for this. It’s just that there is
something between the current codebase and that of 23 April 2018 (i.e.
our latest snapshot) that is different and that results in the issue
this commit was trying to address. In fact, there is another issue,
which is that the Simulation Experiment view flashes a bit when
switching to it and there are several graph panels, which results in
the N62 SED-ML file being considered modified while it clearly
shouldn’t be (and isn’t using the 2018-04-23 snapshot). So, we need to
figure out what we did that causes those issues…