Added a type definition for QBoolSet, QIntSet and QDoubleSet.
So we are in sync with our QBoolList, QIntList and QDoubleList types,
and also the fact that a QSet can be useful when order is not important
and we call the `contains()` method many times. Indeed, QList has a
runtime complexity of O(n) while QSet only has a runtime complexity of
O(1).