Revert "Central widget: make sure that our views are hidden before actually showing the one we are after (#1683)."
This reverts commit 23b6d13b07e6d01aa2e1a46b110b8a7d98779006.

Indeed, we shouldn’t have to hide the views tabs first before showing
the one we are after. It’s just that our Simulation Experiment view is
such that one of its side effects is a GUI glitch that shows both the
Editing and the Simulation views tabs for a split second when switching
files, with the first one using the Simulation Experiment view and the
second using, say, the CellML Text view.