Central widget: get our loaded GUI plugins to update themselves in CentralWidget::updateGui() rather than MainWindow::updateGui().
This work needs a bit more refactoring, but the idea is to prevent a
GUI glitch when opening a remote file. Indeed, such a file would, in a
text-based view, have a reddish background. However, that background
would not always appear and when it would, it would only do so after a
split second, which is clearly not ideal. This didn’t use to happen,
but is the indirect result of our use of the busy widget (see the notes
in commit 5ed445d).