Help window: make sure that we can directly see images (#1544).
Indeed, when we have a page with an image on its own (e.g. the GUI page
in the user documentation), then Sphinx will generate something like
<a><img></a>, which means that we can click on the image to get the
image itself (as opposed to a web page with an image in it). Until now,
our Help window only supported "proper" web pages, not images or
anything else. Now, it should support anything that a Web browser
supports.