Graph Panel widget: use 2 as our default graph line width.
We set it (back) to 1 in commit dddabb7 since a bigger width means that
it will be more time consuming to draw and that we didn’t want to use a
big width by default (this is particularly noticeable using the
noble_model_1962 3.zip file in issue #1806). However, a width of 1
looks really thin while a width of 2 looks just right. Also, it’s not
everyday that we run simulations with 1,000,000 data points, so I guess
it’s fine to go with a default line width of 2.