Simulation Experiment view: don't do anything if a file has been modified.
Until now, if a CellML file had been modified (outside of OpenCOR) and
we hadn’t reloaded it, then we enabled our Reset Model Parameters
action while there is no need for this. This action should only ever be
enabled based on a simulation’s state, not a file’s state.