Graph Panel widget: handle the case where we want to optimise a number that is already optimised.
The case at hand was when we had a pMin value of -0.3 and a minorStep
value of 0.05. You would expect the new value (linear) value of pMin
still to be -0.3, but in our particular case it was -0.35 because
qFloor(pMin/minorStep) was giving us -7 rather than -6 as we would have
expected.

So, to go around this problem, we do a fuzzy null check of
pMin/minorStep against its integer representation…