Main window: removed `&` from different menu items.
Hopefully, they don’t get used. Indeed, the reason for removing those
`&` is that it makes it much faster to create snapshots on Windows by
pressing Alt+PrtScn. However, with `&` in our main menu items, it meant
that the characters used as a shortcut (i.e. Alt+Char) were underlined,
which is not what we want…