doNothing(): make it slower and real time when it comes to its number of loops.
We used to pass a value of 100*delay to doNothing(), but it’s better to
slow doNothing() itself and we now do so by slowing it down by 1,000,
i.e. 10 times slower than it used to be (to account for faster
computers). Also, we now pass a pointer to doNothing() so that it can
adapt to real-time changes to our delay.