File Organiser window: slightly reworked the loading/saving of settings.
Indeed, we used to track the index of an item’s parent item, but we
don’t actually need to do that. This also means that we had to put a
safeguard in FileOrganiserWindowModel::decodeHierarchyData() since
otherwise OpenCOR would crash if it was to retrieve some settings that
were stored using the old format. In the end, that safeguard is a good
thing to have no matter what, so that’s fine.