Data store: improved our DataStoreImportData constructor (#1845).
Indeed, if adding variables was to fail, we were not able to remove
them since the call to DataStore::addVariables() didn’t return
properly. So, now, we add one variable at a time and keep track of the
ones that we were able to add, in case we would need to remove them.