Some minor clarifications.
Re the data store importer/exporter, if we were not to pass
Qt::UniqueConnection, then if we want to import n signals, then the
first time we would do that, we would indeed get n signals imported,
but the second time we would get 2n of them, the third time 3n, etc.
hence the need to pass Qt::UniqueConnection.