libSEDML: removed our use of an explicit release tag (#2042).
The release tag used to be `0.4.4`, which required our use of an
explicit release tag since by default we would expect the release tag
to be `v0.4.4`. However, a new `v0.4.4` release tag has been created on
GitHub, hence now we don’t need to specify an explicit release tag.